refactor: define `restart_after` as a `float`

+++ tests/test_agents/base.py
@@ -353,7 +353,7 @@ class BaseTestEnd2EndAgentCatchup(BaseTestEnd2End):
     # mandatory argument
     stop_string: str
 
-    restart_after: int = 60
+    restart_after: float = 60
     wait_before_stop: int = 15
 
     def setup(self) -> None:
